avatar/tagger/requirements.txt
pillow>=9.0.0
piexif>=1.1.3
opencv-python
huggingface-hub
gradio
numpy
pandas
onnxruntime-gpu
tqdm